Remove skipping of multiline comments
Currently, multiline comments are assumed to be comments that begin with
`/*` and end with `*/`. This assumption itself is flawed as such a
comment can be wholly in a single line. If such a situation is
encountered, then the `in_multiline_comment` flag will never be reset to
False resulting in many lines not being processed at all.

The other problem with the current implementation is that it assumes
that if a line starts with `/*`, then the whole line must be a comment
and is ignored. This is not necessarily true.

It would be good to handle `/* comment */` style of comments, but this
is no meab feat. Until then such comments are treated as valid code.
This has the drawback that import statements in comments (e.g. in usage
examples) may get trimmed or symbols found in comments may be treated as
symbols seen. But these cases generally do less damage than skipping
code assuming that it is just a comment.